requirements.txt
aiohttp             # for async http requests
asyncio             # for async i/o n
pyrofork            # updated alternative to pyrogram
python-dotenv       # used for reading .env files
pytz                # time-zone setup
tgcrypto            # required for faster connection
uvloop              # for improved async networking
vanilite            # for string localisation

bs4
markdown

# requests

# serial: asyncio* > aiohttp > python-dotenv > pytz > tgcrypto* > pyrofork > uvloop > vanilite?